Added the ability to parse polymorphic server response - Bug fix (#14)
Add ability for gson to parse polymorphic server responses

If we intend to parse polymorphic responses from the server - this is how we do it. As of for now, this is the simplest way to do it.
Every polymorphic class must have a common field (in this case - classType) and when we parse we must use a Gson with the implemented AdapterFactory (see code). If we use a Gson with this AdapterFactory that doesn't have the classType attribute - then the program crashes. Hence the special Gson should be created only when needed.

Till now we printed the json string to the Web GUI. The json string is worthless unless we manage to parse it to a concrete type.

The alternative to this code is to use only one type of response with no polymorphism.